% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcKNNgraph_pearson}
\alias{calcKNNgraph_pearson}
\title{Calculate k Nearest Neighbors from Pearson distance metric}
\usage{
calcKNNgraph_pearson(x, k = 1L)
}
\arguments{
\item{x}{An m x n numeric matrix}

\item{k}{The number of nearest neighbors to return
per sample}
}
\value{
An m x k matrix of indicies 1...k of the
nearest neighbors for the specified row based on
Pearson distance.
}
\description{
Each distance metric has its own function for speed/efficiency
This takes a sample X feature matrix and returns
a matrix of k nearest neighbors. This is the one for Pearson.
}

