% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{methods}
\name{export10X}
\alias{export10X}
\alias{export10X,ANY,character-method}
\title{Export data into three files similar to that of 10X pre-processing}
\usage{
export10X(obj, folder)

\S4method{export10X}{ANY,character}(obj, folder)
}
\arguments{
\item{obj}{A sparse matrix object to be exported}

\item{folder}{Which folder should the data be exported to}
}
\value{
Nothing; just text saying that it was successful or not. 
This function wil write files to the disk though.
}
\description{
\code{export10X} takes a sparse matrix and then
exports the object into the files that 10X data typically comes in.
}
\examples{
file1 <- paste(system.file("extdata", package = "indRop"),
"small1.counts.tsv", sep = "/")
file2 <- paste(system.file("extdata", package = "indRop"),
"small2.counts.tsv", sep = "/")
dat <- readIndrop(c(file1,file2))


}
