% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{methods}
\name{readIndrop}
\alias{readIndrop}
\alias{readIndrop,character-method}
\title{Import indRop processed data}
\usage{
readIndrop(paths, whichFile = FALSE)

\S4method{readIndrop}{character}(paths, whichFile = FALSE)
}
\arguments{
\item{paths}{File paths (one or more) to .tsv files generated by indrops analyses}

\item{whichFile}{= FALSE Vector that corresponds to which files each cell came from}
}
\value{
Minimally, a sparse matrix that can be used for Seurat. May be contained
inside an S3 object if other parameters are set to TRUE
}
\description{
\code{readIndrop} takes a (vector of) file-path(s) that point to .tsv files
that were the output of indrops pre-processing pipeline (python package).
}
\examples{

}
