% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnGraph.R
\docType{methods}
\name{spRing}
\alias{spRing}
\alias{spRing,ANY,numeric-method}
\title{Execute the spRing algorithm}
\usage{
spRing(x, k = 10, method = "pearson", knngraph = FALSE, igraph = FALSE)

\S4method{spRing}{ANY,numeric}(x, k = 10, method = "pearson",
  knngraph = FALSE, igraph = FALSE)
}
\arguments{
\item{x}{Data matrix}

\item{k}{= 10}

\item{method}{= "pearson" The distance metric for the data.}

\item{knngraph}{= FALSE Return knn graph generated by the data?}

\item{igraph}{= FALSE Return igraph network associated with knn graph?}
}
\value{
One or more outputs. See above
}
\description{
\code{spRing} takes a matrix of samples X features and computes a
per-sample distance vector before selecting the k nearest neighbors.
This is done iteratively in \code{Rcpp} to minimize the memory
footprint while optimizing speed. Force directed layout is implemented
from \code{igraphs layout_with_drl} function. If either/both knngraph
or igraph is TRUE, this function will return an S3 object (list) with
the coordinates and either/both of these objects. Otherwise, a
two dimensional vector specifying the coordinates will be returned.
}
\examples{
x <- matrix(rnorm(100), nrow = 20) # 20 samples
coords <- spRing(x, k = 2, method = "euclidean")
# plot(coords)
}

